updated strategy for retreiving and checking session cookie

+++ apps/web/src/actions/campaign/delete-many-campaigns.ts
@@ -1,13 +1,15 @@
 "use server"
 
 import { routes } from "@repo/routes"
-import { DeleteManyCampaignsDto } from "@repo/types/campaign"
+import { SESSION_COOKIE } from "@repo/types/auth"
+import { type DeleteManyCampaignsDto } from "@repo/types/campaign"
 
-import { sessionMiddleware } from "~/actions/utils"
+import { getCookie } from "~/lib/cookies"
 import { getBackendPrivateUrl } from "~/lib/url"
 
 const deleteManyCampaigns = async (organizationId: string, dto: DeleteManyCampaignsDto): Promise<void> => {
-	const sessionToken = await sessionMiddleware()
+	const sessionToken = await getCookie(SESSION_COOKIE)
+	if (!sessionToken) throw new Error("Unauthorized")
 
 	const backendUrl = getBackendPrivateUrl()
 	const res = await fetch(`${backendUrl}${routes.backend.DELETE_MANY_CAMPAIGNS(organizationId)}`, {
